refac: Change integer types to int64 for pagination
- Refactor pagination parameters in various queries to use int64
  instead of int32 for better compatibility with larger datasets.
- Modify SQL queries to handle UUIDs instead of integers for
  brand and collection filtering.
- Remove unused SQL queries related to category sort order and
  product images.

+++ server/internal/db/repository/order.sql.go
@@ -290,8 +290,8 @@ OFFSET $3
 
 type ListOrderItemsParams struct {
 	OrderID uuid.UUID `json:"order_id"`
-	Limit   int32     `json:"limit"`
-	Offset  int32     `json:"offset"`
+	Limit   int64     `json:"limit"`
+	Offset  int64     `json:"offset"`
 }
 
 func (q *Queries) ListOrderItems(ctx context.Context, arg ListOrderItemsParams) ([]OrderItem, error) {
@@ -346,8 +346,8 @@ OFFSET $2
 `
 
 type ListOrdersParams struct {
-	Limit      int32              `json:"limit"`
-	Offset     int32              `json:"offset"`
+	Limit      int64              `json:"limit"`
+	Offset     int64              `json:"offset"`
 	CustomerID pgtype.UUID        `json:"customer_id"`
 	Status     NullOrderStatus    `json:"status"`
 	StartDate  pgtype.Timestamptz `json:"start_date"`
